separated king.cpp into a header and the corresponding file.
